// RatelCloudServices/AdminPortalAreo/src/app/authentication/models/user.ts
export class User {
    uid: string;
    uname: string;
    pwd: string;
    email: string;
    code: string;
    token: string;
    key: string;
    avatar: string;

    constructor() {
        this.uid = null;
        this.uname = null;
        this.pwd = null;
        this.email = null;
        this.code = null;
        this.token = null;
        this.key = null;
        this.avatar = null;
    }
}